pyproject.toml
[tool.poetry]
name = "mypydot"
version = "2022.1.1"
description = "Python package to manage your dotfiles"
authors = ["Andrés Ortiz <andrs.ortizl@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/andres-ortizl/mypydot"
repository = "https://github.com/andres-ortizl/mypydot"
keywords = ["mypydot", "dotfiles", "python"]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Environment :: Console",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux"
]


include = [
    "LICENSE",
]
[tool.poetry.dependencies]
python = "^3.11.1"
PyYAML = "6.0"
emoji = "^2.2.0"
PyTermGUI = "^7.3.0"


[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
pytest-cov = "^4.0.0"
coverage-badge = "^1.1.0"
pre-commit = "^2.21.0"
pyte = "^0.8.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mypydot = 'mypydot.src.main:entry_point'
